Split library into package and adapt examples+doc
