+++ README.rst
@@ -66,11 +66,11 @@ Matrix:
 
 .. code:: python
 
-    import adafruit_is31fl3731
+    from adafruit_is31fl3731 import Matrix
     import board
     import busio
     with busio.I2C(board.SCL, board.SDA) as i2c:
-        display = adafruit_is31fl3731.Matrix(i2c)
+        display = Matrix(i2c)
         display.fill(127)
 
 
@@ -78,11 +78,11 @@ Charlie Wing:
 
 .. code:: python
 
-    import adafruit_is31fl3731
+    from adafruit_is31fl3731.CharlieWing import CharlieWing
     import board
     import busio
     with busio.I2C(board.SCL, board.SDA) as i2c:
-        display = adafruit_is31fl3731.CharlieWing(i2c)
+        display = CharlieWing(i2c)
         display.fill(127)
 
         # Turn off pixel 4,4, change its brightness and turn it back on
