+++ adafruit_is31fl3731/CharlieBonnet.py
@@ -1,0 +1,64 @@
+# The MIT License (MIT)
+#
+# Copyright (c) 2017 Tony DiCola
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+"""
+`adafruit_is31fl3731`
+====================================================
+
+CircuitPython driver for the IS31FL3731 charlieplex IC.
+
+
+* Author(s): Tony DiCola, Melissa LeBlanc-Williams
+
+Implementation Notes
+--------------------
+
+**Hardware:**
+
+* `Adafruit 16x9 Charlieplexed PWM LED Matrix Driver - IS31FL3731
+  <https://www.adafruit.com/product/2946>`_
+
+* `Adafruit 15x7 CharliePlex LED Matrix Display FeatherWings
+  <https://www.adafruit.com/product/2965>`_
+
+**Software and Dependencies:**
+
+* Adafruit CircuitPython firmware (2.2.0+) for the ESP8622 and M0-based boards:
+  https://github.com/adafruit/circuitpython/releases
+"""
+
+# imports
+from . import Matrix
+
+
+class CharlieBonnet(Matrix):
+    """Supports the Charlieplexed bonnet"""
+
+    width = 16
+    height = 8
+
+    @staticmethod
+    def pixel_addr(x, y):
+        """Calulate the offset into the device array for x,y pixel"""
+        if x >= 8:
+            return (x - 6) * 16 - (y + 1)
+        return (x + 1) * 16 + (7 - y)
