+++ adafruit_is31fl3731/__init__.py
@@ -364,139 +364,3 @@ class Matrix:
         for x in range(self.width):  # yes this double loop is slow,
             for y in range(self.height):  #  but these displays are small!
                 self.pixel(x, y, pixels[(x, y)], blink=blink, frame=frame)
-
-
-class CharlieWing(Matrix):
-    """Supports the Charlieplexed feather wing
-    """
-
-    width = 15
-    height = 7
-
-    @staticmethod
-    def pixel_addr(x, y):
-        """Calulate the offset into the device array for x,y pixel
-        """
-        if x > 7:
-            x = 15 - x
-            y += 8
-        else:
-            y = 7 - y
-        return x * 16 + y
-
-
-class CharlieBonnet(Matrix):
-    """Supports the Charlieplexed bonnet"""
-
-    width = 16
-    height = 8
-
-    @staticmethod
-    def pixel_addr(x, y):
-        """Calulate the offset into the device array for x,y pixel"""
-        if x >= 8:
-            return (x - 6) * 16 - (y + 1)
-        return (x + 1) * 16 + (7 - y)
-
-
-class ScrollPhatHD(Matrix):
-    """Supports the Scroll pHAT HD by Pimoroni"""
-
-    width = 17
-    height = 7
-
-    @staticmethod
-    def pixel_addr(x, y):
-        """Translate an x,y coordinate to a pixel index."""
-        if x <= 8:
-            x = 8 - x
-            y = 6 - y
-        else:
-            x = x - 8
-            y = y - 8
-        return x * 16 + y
-
-
-class LedShim(Matrix):
-    """Supports the LED SHIM by Pimoroni"""
-
-    width = 28
-    height = 3
-
-    def __init__(self, i2c, address=0x75):
-        super().__init__(i2c, address)
-
-    # pylint: disable-msg=too-many-arguments
-    def pixelrgb(self, x, r, g, b, blink=None, frame=None):
-        """
-        Blink or brightness for x-pixel
-
-        :param x: horizontal pixel position
-        :param r: red brightness value 0->255
-        :param g: green brightness value 0->255
-        :param b: blue brightness value 0->255
-        :param blink: True to blink
-        :param frame: the frame to set the pixel
-        """
-        super().pixel(x, 0, r, blink, frame)
-        super().pixel(x, 1, g, blink, frame)
-        super().pixel(x, 2, b, blink, frame)
-
-        # pylint: disable=inconsistent-return-statements
-        # pylint: disable=too-many-return-statements
-        # pylint: disable=too-many-branches
-
-    @staticmethod
-    def pixel_addr(x, y):
-        """Translate an x,y coordinate to a pixel index."""
-        if y == 0:
-            if x < 7:
-                return 118 - x
-            if x < 15:
-                return 141 - x
-            if x < 21:
-                return 106 + x
-            if x == 21:
-                return 15
-            return x - 14
-
-        if y == 1:
-            if x < 2:
-                return 69 - x
-            if x < 7:
-                return 86 - x
-            if x < 12:
-                return 28 - x
-            if x < 14:
-                return 45 - x
-            if x == 14:
-                return 47
-            if x == 15:
-                return 41
-            if x < 21:
-                return x + 9
-            if x == 21:
-                return 95
-            if x < 26:
-                return x + 67
-            return x + 50
-
-        if x == 0:
-            return 85
-        if x < 7:
-            return 102 - x
-        if x < 11:
-            return 44 - x
-        if x < 14:
-            return 61 - x
-        if x == 14:
-            return 63
-        if x < 17:
-            return 42 + x
-        if x < 21:
-            return x + 25
-        if x == 21:
-            return 111
-        if x < 27:
-            return x + 83
-        return 93
