+++ examples/is31fl3731_blink_example.py
@@ -1,20 +1,29 @@
-import busio
 import board
-import adafruit_is31fl3731
+import busio
+
+# uncomment next line if you are using Feather CharlieWing LED 15 x 7
+from adafruit_is31fl3731.CharlieWing import CharlieWing
+
+# uncomment next line if you are using Adafruit 16x9 Charlieplexed PWM LED Matrix
+# from adafruit_is31fl3731 import Matrix
+# uncomment next line if you are using Adafruit 16x8 Charlieplexed Bonnet
+# from adafruit_is31fl3731.CharlieBonnet import CharlieBonnet
+# uncomment next line if you are using Pimoroni Scroll Phat HD LED 17 x 7
+# from adafruit_is31fl3731.ScrollPhatHD import ScrollPhatHD
 
 i2c = busio.I2C(board.SCL, board.SDA)
 
 # array pattern in bits; top row-> bottom row, 8 bits in each row
 an_arrow = bytearray((0x08, 0x0C, 0xFE, 0xFF, 0xFE, 0x0C, 0x08, 0x00, 0x00))
 
-# initial display using Feather CharlieWing LED 15 x 7
-display = adafruit_is31fl3731.CharlieWing(i2c)
+# uncomment next line if you are using Feather CharlieWing LED 15 x 7
+display = CharlieWing(i2c)
 # uncomment next line if you are using Adafruit 16x9 Charlieplexed PWM LED Matrix
-# display = adafruit_is31fl3731.Matrix(i2c)
-# uncomment line if you are using Adafruit 16x9 Charlieplexed PWM LED Matrix
-# display = adafruit_is31fl3731.CharlieBonnet(i2c)
-# initial display using Pimoroni Scroll Phat HD LED 17 x 7
-# display = adafruit_is31fl3731.ScrollPhatHD(i2c)
+# display = Matrix(i2c)
+# uncomment next line if you are using Adafruit 16x8 Charlieplexed Bonnet
+# display = CharlieBonnet(i2c)
+# uncomment next line if you are using Pimoroni Scroll Phat HD LED 17 x 7
+# display = ScrollPhatHD(i2c)
 
 # first load the frame with the arrows; moves the an_arrow to the right in each
 # frame
