+++ examples/is31fl3731_wave_example.py
@@ -1,6 +1,15 @@
 import board
 import busio
-import adafruit_is31fl3731
+
+# uncomment next line if you are using Feather CharlieWing LED 15 x 7
+from adafruit_is31fl3731.CharlieWing import CharlieWing
+
+# uncomment next line if you are using Adafruit 16x9 Charlieplexed PWM LED Matrix
+# from adafruit_is31fl3731 import Matrix
+# uncomment next line if you are using Adafruit 16x8 Charlieplexed Bonnet
+# from adafruit_is31fl3731.CharlieBonnet import CharlieBonnet
+# uncomment next line if you are using Pimoroni Scroll Phat HD LED 17 x 7
+# from adafruit_is31fl3731.ScrollPhatHD import ScrollPhatHD
 
 i2c = busio.I2C(board.SCL, board.SDA)
 
@@ -33,14 +42,14 @@ sweep = [
 
 frame = 0
 
-# initialize display using Feather CharlieWing LED 15 x 7
-display = adafruit_is31fl3731.CharlieWing(i2c)
+# uncomment next line if you are using Feather CharlieWing LED 15 x 7
+display = CharlieWing(i2c)
 # uncomment next line if you are using Adafruit 16x9 Charlieplexed PWM LED Matrix
-# display = adafruit_is31fl3731.Matrix(i2c)
+# display = Matrix(i2c)
 # uncomment next line if you are using Adafruit 16x8 Charlieplexed Bonnet
-# display = adafruit_is31fl3731.CharlieBonnet(i2c)
-# initial display using Pimoroni Scroll Phat HD LED 17 x 7
-# display = adafruit_is31fl3731.ScrollPhatHD(i2c)
+# display = CharlieBonnet(i2c)
+# uncomment next line if you are using Pimoroni Scroll Phat HD LED 17 x 7
+# display = ScrollPhatHD(i2c)
 
 while True:
     for incr in range(24):
